moved helper functions to utils folder
