+++ client/src/components/Schedule/ScheduleBox.tsx
@@ -1,7 +1,7 @@
 import { TimeSlot as TimeSlotType } from '../../types';
 import TimeSlot from './TimeSlot';
 import { useGetScheduleQuery } from '../../redux/services/schedule/scheduleService';
-
+import { calculateHeight, calculateDistanceFromTop } from '../../utils/scheduleUtils';
 type Props = {
   timeSlots: TimeSlotType[] | undefined;
 };
@@ -10,51 +10,6 @@ const ScheduleBox = ({ timeSlots }: Props) => {
   const { data, isFetching } = useGetScheduleQuery('schedule', {
     pollingInterval: 900000,
   });
-
-  const convertTimeToMinutes = (time: string) => {
-    const lastTwoChars: string = time.slice(time.length - 2, time.length);
-    time = time.slice(0, time.indexOf(' '));
-    const [hour, minutes] = time.split(':');
-
-    let hourNumber: number = parseInt(hour);
-    const minutesNumber: number = parseInt(minutes);
-
-    if (lastTwoChars === 'PM' && hourNumber < 12) {
-      hourNumber += 12;
-    }
-
-    return [hourNumber, minutesNumber];
-  };
-
-  const calculateHeight = (startTime: string, endTime: string) => {
-    const [startHour, startMinutes] = convertTimeToMinutes(startTime);
-    const [endHour, endMinutes] = convertTimeToMinutes(endTime);
-
-    let hours: number = endHour - startHour;
-    let minutes: number = endMinutes - startMinutes;
-
-    if (minutes < 0) {
-      minutes += 60;
-      hours -= 1;
-    }
-
-    const totalMinutes: number = hours * 60 + minutes;
-    const height: number = (totalMinutes * 72) / 60;
-
-    return height.toString();
-  };
-
-  const calculateMinutesFromTop = (time: string) => {
-    const [hour, minutes] = convertTimeToMinutes(time);
-    const minutesFromTop: number = (hour - 7) * 60 + minutes;
-    return minutesFromTop;
-  };
-
-  const calculateDistanceFromTop = (startTime: string) => {
-    const minutes: number = calculateMinutesFromTop(startTime);
-    const distanceFromTop: number = (minutes * 72) / 60;
-    return distanceFromTop.toString();
-  };
 
   if (!isFetching) {
   }
@@ -91,9 +46,7 @@ const ScheduleBox = ({ timeSlots }: Props) => {
           <hr className="absolute top-[864px] w-full border-dotted bg-gray-400 dark:bg-gray-900" />
           <hr className="absolute top-[936px] w-full border-dotted bg-gray-400 dark:bg-gray-900" />
           <div className="relative mx-2 h-[1008px] w-1/5">
-            <h2 className="absolute -inset-8 text-center text-lg font-medium capitalize">
-              {window.innerWidth < 535 ? 'm' : 'monday'}
-            </h2>
+            <h2 className="absolute -inset-8 text-center text-lg font-medium capitalize">{window.innerWidth < 535 ? 'm' : 'monday'}</h2>
             {timeSlots
               ? timeSlots
                   .filter((timeSlot: TimeSlotType) => timeSlot.days.monday)
@@ -102,10 +55,7 @@ const ScheduleBox = ({ timeSlots }: Props) => {
                       key={timeSlot._id}
                       id={timeSlot._id}
                       top={calculateDistanceFromTop(timeSlot.startTime)}
-                      height={calculateHeight(
-                        timeSlot.startTime,
-                        timeSlot.endTime
-                      )}
+                      height={calculateHeight(timeSlot.startTime, timeSlot.endTime)}
                       title={timeSlot.title}
                       startTime={timeSlot.startTime}
                       endTime={timeSlot.endTime}
@@ -123,10 +73,7 @@ const ScheduleBox = ({ timeSlots }: Props) => {
                       key={timeSlot._id}
                       id={timeSlot._id}
                       top={calculateDistanceFromTop(timeSlot.startTime)}
-                      height={calculateHeight(
-                        timeSlot.startTime,
-                        timeSlot.endTime
-                      )}
+                      height={calculateHeight(timeSlot.startTime, timeSlot.endTime)}
                       title={timeSlot.title}
                       startTime={timeSlot.startTime}
                       endTime={timeSlot.endTime}
@@ -138,9 +85,7 @@ const ScheduleBox = ({ timeSlots }: Props) => {
                   ))}
           </div>
           <div className="relative mx-2 h-[1008px] w-1/5">
-            <h2 className="absolute -inset-8 text-center text-lg font-medium capitalize">
-              {window.innerWidth < 535 ? 't' : 'tuesday'}
-            </h2>
+            <h2 className="absolute -inset-8 text-center text-lg font-medium capitalize">{window.innerWidth < 535 ? 't' : 'tuesday'}</h2>
             {timeSlots
               ? timeSlots
                   .filter((timeSlot: TimeSlotType) => timeSlot.days.tuesday)
@@ -149,10 +94,7 @@ const ScheduleBox = ({ timeSlots }: Props) => {
                       key={timeSlot._id}
                       id={timeSlot._id}
                       top={calculateDistanceFromTop(timeSlot.startTime)}
-                      height={calculateHeight(
-                        timeSlot.startTime,
-                        timeSlot.endTime
-                      )}
+                      height={calculateHeight(timeSlot.startTime, timeSlot.endTime)}
                       title={timeSlot.title}
                       startTime={timeSlot.startTime}
                       endTime={timeSlot.endTime}
@@ -170,10 +112,7 @@ const ScheduleBox = ({ timeSlots }: Props) => {
                       key={timeSlot._id}
                       id={timeSlot._id}
                       top={calculateDistanceFromTop(timeSlot.startTime)}
-                      height={calculateHeight(
-                        timeSlot.startTime,
-                        timeSlot.endTime
-                      )}
+                      height={calculateHeight(timeSlot.startTime, timeSlot.endTime)}
                       title={timeSlot.title}
                       startTime={timeSlot.startTime}
                       endTime={timeSlot.endTime}
@@ -185,9 +124,7 @@ const ScheduleBox = ({ timeSlots }: Props) => {
                   ))}
           </div>
           <div className="relative mx-2 h-[1008px] w-1/5">
-            <h2 className="absolute -inset-8 text-center text-lg font-medium capitalize">
-              {window.innerWidth < 535 ? 'w' : 'wednesday'}
-            </h2>
+            <h2 className="absolute -inset-8 text-center text-lg font-medium capitalize">{window.innerWidth < 535 ? 'w' : 'wednesday'}</h2>
             {timeSlots
               ? timeSlots
                   .filter((timeSlot: TimeSlotType) => timeSlot.days.wednesday)
@@ -196,10 +133,7 @@ const ScheduleBox = ({ timeSlots }: Props) => {
                       key={timeSlot._id}
                       id={timeSlot._id}
                       top={calculateDistanceFromTop(timeSlot.startTime)}
-                      height={calculateHeight(
-                        timeSlot.startTime,
-                        timeSlot.endTime
-                      )}
+                      height={calculateHeight(timeSlot.startTime, timeSlot.endTime)}
                       title={timeSlot.title}
                       startTime={timeSlot.startTime}
                       endTime={timeSlot.endTime}
@@ -217,10 +151,7 @@ const ScheduleBox = ({ timeSlots }: Props) => {
                       key={timeSlot._id}
                       id={timeSlot._id}
                       top={calculateDistanceFromTop(timeSlot.startTime)}
-                      height={calculateHeight(
-                        timeSlot.startTime,
-                        timeSlot.endTime
-                      )}
+                      height={calculateHeight(timeSlot.startTime, timeSlot.endTime)}
                       title={timeSlot.title}
                       startTime={timeSlot.startTime}
                       endTime={timeSlot.endTime}
@@ -232,9 +163,7 @@ const ScheduleBox = ({ timeSlots }: Props) => {
                   ))}
           </div>
           <div className="relative mx-2 h-[1008px] w-1/5">
-            <h2 className="absolute -inset-8 text-center text-lg font-medium capitalize">
-              {window.innerWidth < 535 ? 'th' : 'thursday'}
-            </h2>
+            <h2 className="absolute -inset-8 text-center text-lg font-medium capitalize">{window.innerWidth < 535 ? 'th' : 'thursday'}</h2>
             {timeSlots
               ? timeSlots
                   .filter((timeSlot: TimeSlotType) => timeSlot.days.thursday)
@@ -243,10 +172,7 @@ const ScheduleBox = ({ timeSlots }: Props) => {
                       key={timeSlot._id}
                       id={timeSlot._id}
                       top={calculateDistanceFromTop(timeSlot.startTime)}
-                      height={calculateHeight(
-                        timeSlot.startTime,
-                        timeSlot.endTime
-                      )}
+                      height={calculateHeight(timeSlot.startTime, timeSlot.endTime)}
                       title={timeSlot.title}
                       startTime={timeSlot.startTime}
                       endTime={timeSlot.endTime}
@@ -264,10 +190,7 @@ const ScheduleBox = ({ timeSlots }: Props) => {
                       key={timeSlot._id}
                       id={timeSlot._id}
                       top={calculateDistanceFromTop(timeSlot.startTime)}
-                      height={calculateHeight(
-                        timeSlot.startTime,
-                        timeSlot.endTime
-                      )}
+                      height={calculateHeight(timeSlot.startTime, timeSlot.endTime)}
                       title={timeSlot.title}
                       startTime={timeSlot.startTime}
                       endTime={timeSlot.endTime}
@@ -279,9 +202,7 @@ const ScheduleBox = ({ timeSlots }: Props) => {
                   ))}
           </div>
           <div className="relative mx-2 h-[1008px] w-1/5">
-            <h2 className="absolute -inset-8 text-center text-lg font-medium capitalize">
-              {window.innerWidth < 535 ? 'f' : 'friday'}
-            </h2>
+            <h2 className="absolute -inset-8 text-center text-lg font-medium capitalize">{window.innerWidth < 535 ? 'f' : 'friday'}</h2>
             {timeSlots
               ? timeSlots
                   .filter((timeSlot: TimeSlotType) => timeSlot.days.friday)
@@ -290,10 +211,7 @@ const ScheduleBox = ({ timeSlots }: Props) => {
                       key={timeSlot._id}
                       id={timeSlot._id}
                       top={calculateDistanceFromTop(timeSlot.startTime)}
-                      height={calculateHeight(
-                        timeSlot.startTime,
-                        timeSlot.endTime
-                      )}
+                      height={calculateHeight(timeSlot.startTime, timeSlot.endTime)}
                       title={timeSlot.title}
                       startTime={timeSlot.startTime}
                       endTime={timeSlot.endTime}
@@ -311,10 +229,7 @@ const ScheduleBox = ({ timeSlots }: Props) => {
                       key={timeSlot._id}
                       id={timeSlot._id}
                       top={calculateDistanceFromTop(timeSlot.startTime)}
-                      height={calculateHeight(
-                        timeSlot.startTime,
-                        timeSlot.endTime
-                      )}
+                      height={calculateHeight(timeSlot.startTime, timeSlot.endTime)}
                       title={timeSlot.title}
                       startTime={timeSlot.startTime}
                       endTime={timeSlot.endTime}
