+++ client/src/pages/UserPage.tsx
@@ -5,6 +5,7 @@ import { useParams } from 'react-router-dom';
 import { Spinner } from 'flowbite-react';
 import { calculateAge } from '../utils/functions';
 import { Button } from 'flowbite-react';
+import ScheduleBox from '../components/Schedule/ScheduleBox';
 const UserPage = () => {
   const { userId } = useParams();
   const { data, isLoading } = useGetExternalUserInfoQuery(userId!);
@@ -19,15 +20,15 @@ const UserPage = () => {
   return (
     <>
       {userInfo ? (
-        <div className="h-screen w-screen bg-slate-400 dark:bg-slate-900">
+        <div className="flex h-screen w-screen gap-10 bg-slate-400 dark:bg-slate-900">
           <div className="ml-10 flex h-5/6 w-1/4 flex-col rounded-lg bg-white p-4 dark:bg-slate-800 dark:text-white">
             <div className="flex flex-col items-center gap-3">
               <img src={userInfo.photoURL} className="h-44 w-44 rounded-full border-2" />
-              <h1 className="text-2xl font-bold mb-4">
+              <h1 className="mb-4 text-2xl font-bold">
                 {userInfo.firstName} {userInfo.lastName}
               </h1>
             </div>
-            <div className="w-7/8 flex h-1/2 flex-col gap-4 rounded-lg bg-slate-400 p-5 text-lg font-medium text-white dark:text-black dark:bg-white">
+            <div className="w-7/8 flex h-1/2 flex-col gap-4 rounded-lg bg-slate-400 p-5 text-lg font-medium text-white dark:bg-white dark:text-black">
               <div>
                 <p>Age: {calculateAge(new Date(userInfo.birthday))}</p>
                 <p>School: {userInfo.school}</p>
@@ -50,6 +51,7 @@ const UserPage = () => {
               </div>
             </div>
           </div>
+          <ScheduleBox timeSlots={undefined}/>
         </div>
       ) : (
         <Spinner aria-label="Profile loading spinner" size="xl" />
